Dedupe tests, always store root node in proof
