+++ ethereum/trie/src/main/java/tech/pegasys/pantheon/ethereum/trie/StoredMerklePatriciaTrie.java
@@ -81,7 +81,7 @@ public class StoredMerklePatriciaTrie<K extends BytesValue, V> implements Merkle
   @Override
   public Proof<V> getValueWithProof(final K key) {
     checkNotNull(key);
-    final ProofVisitor<V> proofVisitor = new ProofVisitor<>();
+    final ProofVisitor<V> proofVisitor = new ProofVisitor<>(root);
     final Optional<V> value = root.accept(proofVisitor, bytesToPath(key)).getValue();
     final List<BytesValue> proof =
         proofVisitor.getProof().stream().map(Node::getRlp).collect(Collectors.toList());
